[cop] Add SwaggerRouting and SwaggerDocumentationAdder modules
